[ddk][usb-hid] Remove extra input report ids
The USB HID spec says a report id in a report is optional if the report
has only one id. Use this behavior in fuchsia.

For boot protocol devices, it's not an issue, since they have only one
report id. For report-based devices, we have to parse HID anyway, so
checking whether a report id is present shouldn't pose any real
difficulty.

Change-Id: I79ef285d5c104189b18fc1c4dc9c2bbaaa061308