Add the manifest file itself to the depfile
When we gen.py and change the set of modules, we need to trigger the
mkbootfs rule. Previously, we were missing a dependency on the manifest
file itself, which means we'd miss this build if the manifest file was
all that changed.

It would be more natural to add the manifest file as an input to the gn
action but that makes gn made because it doesn't understand where this
file comes from. Instead, this CL adds it to the depfile for that rule,
which has the sameffect.

Change-Id: Ie622eed530d6df38ba77af69a7d0e772d1202ac3