[docs][dispatcher] Updated mxio_port_bind to use uint64_t
The mxio_port_bind syscall used to use a "uintptr_t" type, but has been
changed to always be 64-bits. Updated documentation to reflect this
fact, and updated dispatcher to avoid casting the key argument to
uintptr_t. The updates to the dispatcher are not necessary, but they
make it clear that the type of 'key' is NOT the size of a pointer.

Change-Id: Ife93b420201ad3d68d92743c22e06f85da3071be