Use clang for host and target builds, remove gcc definitions
We don't need to use gcc for host or target builds on any platforms we
support. This removes the gcc toolchain template and uses the clang
binaries we provide on linux hosts and hardcodes the 'is_clang'
variable, which is tested by various build files we use, to true. We
still use the clang binary from xcode on mac for host binaries.

This also renames the toolchain labels to be less reflective of how they are built and more
reflective of how they are used. The host toolchain is something like

//build/toolchain:host_${host_cpu}

i.e. //build/toolchain:host_x64

and the target toolchain is

//build/toolchain/fuchsia:${target_cpu}

Change-Id: Ic58861a58396c8a8887562ffd54908ffe2bbb932