[make][userspace] Define _ALL_SOURCE for userspace builds
We have defined some commonly used extensions (around thread and lock
creation, in particular) that are feature gated behind the _ALL_SOURCE
macro. Tthese extensions will be used so widely that we do not wish to
litter magenta source with this define. So instead, have the build
system define it for everyone (except libc, but it already disables it).

Change-Id: If1db31c88b040c01fb42efe2ea15560025334e7f