[remoteio] Removed 'mxrio_server' and 'mxrio_handler_create'
These two functions provide mechanisms to launch remote mxio servers.
Their purpose involves waiting for a handle to become readable before
dispatching to the "mxrio_handler" intermediate callback function.

This logic is duplicated with the general purpose dispatcher, which,
when called with "mxrio_handler" as an argument, fulfills the same
function. In fact, the general purpose dispatcher is being used instead
of 'mxrio_server' and 'mxrio_handler_create' elsewhere in Magenta.

Change-Id: I06ec3bdd1a027cf182366e2535f9c6b77358de37