[ulib] [musl] [ldso] Make internal struct match 'struct link_map' again
The upstream musl penchant for repeating layouts rather than using
the proper types (and not even leaving any comments to remind you!)
led me to forget the public 'struct link_map' protocol when I added
the soname field to the internal 'struct dso'.  Move the field so
that the public 'struct link_map' layout is intact.

Change-Id: Ib3651c5b2852e12ea6900422d15d86bd9cc61c7f