[scripts] Add script to symbolize backtrace
This adds a script that reads stdin and echos it back out, symbolizing
any backtraces it recognizes using addr2line. Specifically it looks
for sequences of lines of the form:

bt#1: pc 0x1501da6 sp 0x130fcc0 (minfs.elf,0x1da6)

and parses out the 'minfs.elf' and '0x1da6' strings and passes those
to addr2line -ifpe.  It applies some simple heuristics to try to find
the correct binary to pass to addr2line, the most important being
that if it can't find a match for the first string in the tuple it
tries the first parameter passed to the script. This can be used to
specify the primary executable's filename.

Change-Id: I0f10ffb8d92af491acf7c00f1563ac6e5e50eefe