[kernel][magenta] Add debug syscall to read memory from process
This syscall takes a process id (a koid) rather than a handle
and as such is only useful as a crutch for a debugger.

mx_ssize_t debug_read_memory(
    uint64_t koid, uintptr_t vaddr, mx_size_t len, void* buffer);

It follows the koid to the process and then down to the vmo
that contains the |vaddr| and reads |len| bytes out of it.

Note that is very possible that less than |len| bytes are
read, how many are copied are returned if the return value
is bigger than zero.

Change-Id: I96ef659947e703e250e9208910e8b7aad01063e7