[scripts] Teach symbolize how to find fuchsia binaries
This teaches symbolize how to look in additional locations for
additional binaries and libraries to symbolize and configures the
default x86-64 debug build directory into the default path. The set
of build directories to look for can be modified with the '-b' /
'--build-dir' flag, i.e.

$ ./scripts/symbolize my_awesome_app -b /my/build_dir /my/lib_dir

Also fixes a bug where backtrace lines where the script couldn't find
a source library were deleted from the output instead of encoded as
an unknown symbol.

Change-Id: Ic4c14c755aa2a2ad82d4cf063c3c2c8c3f1c5cd9