Implements atomic updates for the notification-handler service.
This CL is in preparation for a follow-up CL that implements the
default service in Dartlang. Providing atomic updates is necessary
because Cloud Platform Task Queues are not available through JSON
APIs; instead, we use Pub/Sub as our task queue mechanism that
doesn't allow us to control the rate of message sending.

Change-Id: Ia11e44736cdbd2b722bb2124140d29f1e4f50d1e