[utils] Share utils with userspace
Move the majority of the code from the ::utils::* namespace into a
location where it can be used by userspace as well as the kernel.

Arena and fifo_buffer were left behind for now because they both have
syscall dependences which need to be resolved.

Tests also currently remain in kernel-space, they will be hoisted up
to user-land next.

Change-Id: I3ec52d2d2d99f3d99a7be3898d82fa6fdf43c413