[usermode acpi] Introduce acpisvc
acpisvc is the usermode ACPI service.  It will be spawned by devmgr.
acpisvc will maintain a mapping of message pipe handles to ACPI
namespace nodes, with devmgr owning the other half of the message pipe
corresponding to the ACPI system bridge node.

These message pipes can be used for speaking a command/response protocol
to the acpisvc for evaluating a restricted subset of methods in the ACPI
interpreter.  In order to limit the scope of the evaluation privilege,
the protocol only allows a client to request new handles for children of
the node it is bound to.

This diff does not implement the full protocol that we will need, but it
provides enough to allow userspace to initialize the PCIe bus.

Change-Id: I77028a4098b1f318166f27096b52d17f8fe29509