[build] Include dependency DSOs in ABI stubs
Link the ABI stub against the same DSOs the real library uses, so the
stub gets DT_NEEDED entries.  These are not strictly part of the ABI.
But at link time, the linker pays attention to them if the DSO has any
undefined symbols.  In some situations, the presence of the undefined
symbols actually is part of the ABI, so we can't omit them from the
stub.  Since they're there, the linker will want to believe that some
other DSO supplies them.

Change-Id: I59a030957a64033bfb7887fa8afe93e51323f5fa