[musl][cnd_t] Fix cnd_timedwait
The original implementation of cnd_t relied on mtx_t having the same
layout as pthread_mutex_t. However, we have changed mtx_t to be
considerably smaller and simpler, based on mxr_mutex_t. Accordingly,
this patch essentially recreates the cnd_timedwait implementation, with
the internal locking changed. It is essentially the same as the
pthread_cond_timedwait implementation, with some dead code removed due
to the simpler mutex implementation.

Change-Id: If56f3be4547d16a567dfa9cda0954430c126a52c