[devmgr][ddk] device_init() is now void
This can only fail if you pass it a bogus name, which is between unlikely
to impossible depending on the driver.  Instead make it never fail, but
if the name was bogus flag the device so device_add() will fail.

Remove a bunch of code for checking the result of this function in every
driver.

Change-Id: Ib857e838b2f2eca09e56498b68602eeecc4369d9