[ulib] Reduced nested if statements, removed unnecessary gotos
Using a "goto" statment for failure cases where a return value is set
and a resource is freed is not preferable to manually unlocking and
returning. For example:

      r = ERR_VALUE;
      goto fail;
    ...
    fail:
      unlock(resource);
      return r;

vs

      unlock(resource);
      return ERR_VALUE;

The "goto" in this case does not simplify error handling -- it just adds
another control flow path can be taken. Returning immediately on error
(when possible) is clearest from a readability perspective.
Additionally, using a "goto" in this manner combines the paths taken by
"failure case" and "success case" code, which is also confusing without
actually reducing the total LoC count.

Change-Id: Idbd5e1d7c2545f4ce24c6aa3cd5c65b2e9584300