Implements an upload endpoint for modules.
We update the default service to be responsible for receiving,
validating, and uploading modules to cloud storage. Because the
manifest is read in the default service, we can now directly
pass this to the notification-handler service for index updates.

In a follow-up CL, we add authentication to the default service.

Change-Id: I2a6ac02beaf748fa786b6e8672ab8e3e5f49500e