[ulib][uapp] Use atomic types with C11 atomic operations
C11 atomics are only defined on atomic types. C11 atomics implementation
in GCC doesn't seem to be enforcing this property, but the one in Clang
does, and it causes compilation errors.

There are three different ways to specify C11 types:

_Atomic const int * p1;
const _Atomic(int) * p3;
const atomic_int * p2;

While the first two are C11 specific, only the third is compatible with
C++11 atomics; the equivalent types are declared within the std
namespace in the <atomic> header. Since many of Magenta's headers are
used in both C and C++ code, this is the prefer option.

Change-Id: I52d6734c1149e9a28059d786e8645e5a0189ac30