[netsvc][netcp] protocol robustness
Handle retransmit (or not) as appropriate.  In the netboot protocol
family, receiving an identical request (same cookie, arg, data) must
result in an identical response and not otherwise have side effects,
as the host re-transmits requests on timeout.

Similarly, the host ignores all responses that are not entirely
correct to avoid erroring out on an out-of-order resent response or
other stray packet.

Updated netboot_txn() and netfile to obey these rules.

Verified with simulated 5% packet loss.

Change-Id: Ie46515d317734d6dd194a46ab0ab89c053b9e9d9