[udev][usb-xhci] Prevent transfer ring overflows
xhci_queue_transfer() now checks to see if there is room in the transaction ring
and returns ERR_NOT_ENOUGH_BUFFER if there are insufficient TRBs free for a transaction.

iotxns queued from above are now kept on a deferred transaction list until there is
room in the transfer ring to queue it to the hardware.

Note that we do not have to worry about the command queue nor control transactions
queued internally during device enumeration since these are all done synchronously
from a single thread, so there is no chance of overflow.

Change-Id: I572f985243ac8ebb65921d8ff84e3f41637cce2b