[kernel][magenta][socket] socket OOB interface
Also known as the control plane. It allows you to send
a single message out-of-band to the other side and raises
the signal_signaled flag.

The control plane is accessed via read/write with
the flag MX_SOCKET_CONTROL and the message can be up to
MX_SOCKET_CONTROL_MAX_LEN

Change-Id: Ib4de4bd2e4a092c00b5b51c0241ba6dbfde737c6