[devmgr] new experimental openat() device op
This op allows a client of a device to ask the device to create
a child device (providing a sub-path or name) via the openat()
call mechanism.

The default openat() hook returns ERR_NOT_SUPPORTED, as it's a
specialized operation that most devices will not support.

I have some crazy ideas around using this mechanism instead of
things like ptmx for certain kinds of dynamic device creation.

Change-Id: Ie9cbfcb17d3ab9b82d3f3e7d7545b32649516e7d