[runtime][launchpad][userboot] Consolidate stack size computations
This moves the code that knows the ABI requirements for the
stack pointer, and the default user thread stack size, into
a comment place shared by all places that need to know this,
both in kernel and in user mode.

It enforces the ABI-required alignment, rather than assuming that all
stacks are allocated page-aligned.  The stacks for additional user
threads created by musl sometimes put the TLS data above the stack.

Change-Id: I606aa10925760a0a3974fd1546e48ca3830f7cd6