[ddk][usb] Add support for USB composite devices
USB composite devices are devices that support more than one function,
for example an audio device that also have a HID interface for media palyback buttons.
For these devices (identifiable by bDeviceClass == 0 in the device descriptor)
we create child devices representing the interfaces contained in the composite device.

For now we make the top-level device for a composite device unbindable, to ensure that
drivers bind to the appropriate interface instead of the composite device itself.
We may later remove this limitation once we have a way of prioritizing interface devices
over the top-level composite device.

Added ioctl to USB protocol to determine whether a device represents a top-level USB device
or is a child interface device, and updated the lsusb program to use this to filter out
the interfaces.

Also added ioctls for the USB driver support code to retrieve descriptors in an abstract way.
For top-level devices we return the full configuration descriptor list.
For child interfaces we return just the USB interface and its contained descriptors.
Using this we can transition to supporting composite devices without updating any of the drivers.

Change-Id: I12609e729e18e1a4a633def9c33c9bab2ffe2e78