[ulib][magenta][runtime][musl] Give system calls _mx_* names, use them in musl
The mx_* name space is not kosher for the standard C library to use
for its own purpose.  A conforming standard C program can define its
own functions called mx_* and this must not break the C library.

Give all the system call entry points in the vDSO names starting
with _mx_, with the old mx_ names as aliases.  Use only the _mx_
names in runtime and musl.

Change-Id: Id01b57b87c0bb91733e9492159401ff019943822