[dart] Include target name in .packages path, allow specifying source
This puts the target name into the path of the generated .packages
file and allows the invoker of the dart_package template to specify
the location of the source dir. These allow specifying multiple dart
packages in the same BUILD.gn file and allow specifying that the
location of Dart source in a package is in a path in the gendir
instead of next to the BUILD.gn file in the source directory, both of
which make it much easier to depend on generated Dart code.

Change-Id: I13fb0adde06acda2fec017b67e6bb208a1331053