[dart] Allow specifying non-dart dependencies in dart_package
This allows a dart_package() to declare dependencies that are not
themselves Dart packages. This is most useful when depending on a step
that generates Dart code, such as an IDL generator. Non-Dart
dependencies exist in the dependency graph but do not produce any
entries in .packages file and thus don't allow additional imports to
resolve by themselves.

Change-Id: I74fdf9bdac50f546e9c282c2275ee778c42de059