[ktrace] core kernel event tracing facility
This adds the ktrace() tracepoint system, which compiles
out to nothing when WITH_LIB_KTRACE is not included.

Trace packets are accumulated in a buffer and an API is
provided to drain the buffer, start/stop tracing, adjust
trace event group masks to determine which events to
capture, etc.

The default buffer size is 32MB.  It is adjustable with
the commandline option ktrace.bufsize (value in MB).

The default group mask for tracing from boot is all groups.
It is adjustable with ktrace.grpmask.

Change-Id: I491e4303f858021f9bf9600ba22ecb2c763d0e22