[scripts] Add a script for running static analyzer
We have some magenta-specific atatic analyzers that verify
the correct usage of locks such as mutexes and spinlocks. These
analyzers maybe extended in the future. To run the analyzers, we
need to be able to build magenta with Clang. This patch provides
a simple script to run the existing analyzers on magenta.

Change-Id: I608186f34660379e105f984b6550eda1d4a482c3