[scripts] Look up binaries by buildid in ids.txt
This tries to find the path to binaries based on their build id in
files called 'ids.txt' in build dirs in addition to looking up things
by name. This is more reliable and faster than doing a lookup by name
as buildids will uniquely identify the actual binary, assuming a
functional toolchain.

This means symbolize when run without any parameters can symbolize
backtraces for any program. The script does a new search on each new
backtrace, so the script can identify binaries built after the script
started or after it symbolized a previous backtrace, so it's possible
to leave a terminal continuously running loglistener piped into
symbolize and have it symbolize new crashes in an edit->compile->test
loop.

Change-Id: I32e5a3b87ae696d77ea08647d4be69bb68ff9c06