[udev][usb] Emulate root hubs as virtual devices
This allows the usb-hub driver to manage root hubs as if they were USB device based hubs

We now read the XHCI extended configuration data to determine which root hub ports
are USB 2.0 or USB 3.0 ports.
Based on this information, we construct two virtual USB devices to represent the
2.0 and 3.0 parts of the XHCI root hub.
These virtual devices are assigned device_ids just past the maximum number of slots.
The XHCI driver now special cases these device_ids and emulates the control and interrupt
requests sent by the usb-hub driver.
In addition, the usb-hub driver now dynamically allocates its array of USB devices,
based on a new function in the HCI protocol which returns the maximum number of devices

Change-Id: I325e7607f68de6339938fb608a1283a53c7a33bf