[direntry] Added mechanism to cache DOS names
Unicode names are converted to "short names" and possibly "long name"
components. Since distinct Unicode names may convert to the same "short
name", the FAT filesystem may insert a generation number inside the
short name, providing uniqueness.

Problems occur when attempting to re-serialize a dirent with the same
name. If a dirent "foo" is serialized, turned into "FOO~1", and placed
in a directory, then when "foo" is re-serialized, it is converted to
"FOO~2" when it is placed into the directory. This is an unexpected
behavior of generation numbers: the underlying name for direntries
should only be set once during allocation.

To fix this, this patch only converts the Unicode name ("foo") to a
short name ("FOO~1") once, and caches it in-memory, during either:
 - LoadDirent (which is used when reading a dirent from disk), or
 - Serialize (which is used before writing a dirent to disk)

Change-Id: Ida9c91c933536a4f337a0e93e89701f96ca2d9bf