apps/media: use ftl/mtl task and thread support
This CL replaces unsupported task, message loop and thread calls
with ftl/mtl equivalents.

services/audio used a thread pool that's not supported. This has
been replaced by a dedicated thread per output, which is not as
cool. Eventually, the magenta driver model will provide something
along the lines of a completion port to apply a thread pool to
I/O completions.

Change-Id: I9c8378bf4735c7df39051aa8c43c33a75f4d5636