apps/media: remove dependencies on LinearTransform
LinearTransform was removed in a previous CL. TimelineFunction and
TimelineRate do the same things, and LinearTransform was encumbered
with an additional license. This CL replaces the code that used
LinearTransform.

TimelineFunction/Rate and LinearTransform are somewhat confusing to
use all by themselves. Converting from one to the other is very,
very confusing, hence the long explanation here.

LinearTransform and TimelineFunction both represent a linear function
in slope/intercept form and are used for mapping one timeline to
another. LinearTransform has no timeline semantics, referring instead
to 'a' and 'b' domains when talking about intercept and 'numerator'
and 'denominator' when talking about slope. The numerator and b
are in the same domain; likewise, the denominator and a are in the
other domain. So a LinearTransform is a function f(a) where:

    f(a) = (a - a_zero) * numerator / denominator + b_zero

The term 'zero' here refers to intercept. A LinearFunction::Ratio is
just a numerator and denominator (the same terms are used).

LinearTransforms can be applied 'forward' (DoForwardTransform) from
a to b or 'backward' (DoBackwardTransform) from b to a.

TimelineFunction uses timeline terminology so developers don't have
to map its terminology to the application. TimelineFunction refers
to the 'reference' domain and the 'subject' domain. That is, a
TimelineFunction maps a reference timeline into a subject timeline.
Put another way, TimelineFunction produces a subject time value from
a reference time value. The intercept terminology is 'reference_time'
and 'subject_time', and the slope terminology is 'reference_delta'
and 'subject_delta'. So a TimelineFunction is a function f(r) where:

    f(r) = (r - reference_time) * subject_delta / reference_delta +
        subject_time

TimelineRate is just a subject_delta and a reference_delta.

TimelineFunctions can be applied forward (Apply) from reference to
subject or backward (ApplyInverse) from subject to reference. Apply
is also offered as operator(), given that a TimelineFunction is
conceptually a function.

So the terms map like this:

    a/denominator <=> reference
    b/numerator <=> subject

Incidently, LinearTransform::Rate uses the term 'Compose' where
TimelineRate uses the term 'Product' and operator*.

Another difference between LinearTransform and TimelineFunction is
the way that error is handled. Because the slope is given as the
ratio of two integers, applying the slope to a value can produce
rounding error. Ideally, this would never happen, but it is
theoretically possible. The consequences aren't necessarily dire,
but we would like to know (in debug builds) when this happens.
LinearTransform takes the approach of returning a bool that
indicates whether a calculation is precise. This means the calling
sites have to deal with that, and there are a lot of DCHECKs for
that purpose. TimelineFunction checks for error internally and does
the DCHECK there. In production, we would never want e.g. playback
to fail just because of rounding error in these calculations.

Change-Id: I4119b0fccac6ed025565f1e4cd79c2cfa34d48df