[kernel][sched] add new version of thread_preempt, switch magenta's use of thread_yield to that
Previously as threads were woken up in the magenta executive, thread_yield() was usually
invoked, which would move the caller thread to the tail of the run queue and give up the
rest of its time slice. If the system was already loaded down with free running threads,
the responsiveness would suffer greatly.

Now use thread_preempt() instead, which usually puts the caller thread on the head of the
queue and invokes the scheduler. Since everything is (currently) running at the same priority,
this almost always means the current thread will continue to run on the current cpu, and in a
SMP situation one of the other cpus will pick it up. This is probably not exactly the optimal
path either, but the system certainly 'feels' better this way.

Change-Id: Id3cad6469872af7dc7722cd774f38a736f58c164