[unittest] Set the default verbosity to zero.
This is part 1/2.
Part 2 will stop runtests from specifying a verbosity by default.

The goal is to have the same default verbosity, whether the test is
run by runtests or run standalone. This is important as bugs can, and
have, shown up at one verbosity or the other (v=0 or v=1).
Having a test fail, or pass, when run with runtests and then do the
opposite when run standalone leads to extra headscratching,
beyond that required to understand the bug itself, which is something
a test harness should not inflict.

Change-Id: I81f519399b00ccd8c168364c4de757005377654a