[mxsh][runtests] Don't specify a verbosity by default.
And add -q to explicitly specify quiet mode (v=0).

This is part 2/2.
Part 1 changed the default verbosity in unittest to 0 (quiet mode),
which is the runtests default.

The goal is to have the same default verbosity, whether the test is
run by runtests or run standalone. This is important as bugs can, and
have, shown up at one verbosity or the other (v=0 or v=1).
Having a test fail, or pass, when run with runtests and then do the
opposite when run standalone leads to extra headscratching,
beyond that required to understand the bug itself, which is something
a test harness should not inflict.

Change-Id: I6907f1427124cab0820e2d123dbae75c6a9e9d51