[udev][sata] implement IOTXN_SYNC_BEFORE and IOTXN_SYNC_AFTER
If the controller receives a transaction with the flag IOTXN_SYNC_BEFORE it
will pause the port from running more transactions until all in-flight
transactions are complete. If it receives a transaction with flag
IOTXN_SYNC_AFTER it will start that transaction, then pause the port
until all in-flight transactions are complete.

In either case if more transactions without either flag are received after
a transaction with a flag, they will not run until after the sync operation.

Change-Id: I742a9a7eb21bfe363a819e7fa20020be132ab2f1