[intel-ethernet] Ack interrupts at the proper point.
Ack interrupts at the proper point in the IRQ service cycle based on
whether the IRQ was set up to be edge triggered (MSI) or level
triggered (PCI Legacy).

Note: this is temporary.  Once the need for mx_complete_interrupt has
been eliminated, we can get rid of this logic.

Change-Id: Id0363da974c306a7141a683ec6cc183e4d4e94fc