[udev][usb-audio] Add simple USB audio & midi driver
The USB audio specification is complicated and this driver does not
attempt to implement all of it, but it does support typical simple devices.

Devices for each audio/MIDI source and sink are created at
/dev/class/audio and /dev/class/midi.
ioctl can be used to determine whether the device is a source or a sink

For audio devices, ioctls are provided for getting list of supported sample rates
and to set the sample rate.

For now, we only support stereo output with 16 bit samples for output
and mono or stereo 16 bit samples for input.

Digging through the various descriptors to get at the volume level information
is complicated. For now we have a hack in place that sets the volume to 80% for all audio devices.

Change-Id: Ieca889c93c01d79f0f37a820facdfd36faba0a4c