[ktrace][ktracedump] move to variable sized ktrace records
The low 4 bits of the record now indicate the size of the entire
record, in units of sizeof(uint64_t)

ktrace_open(tag) now provides a common path to acquire a record
to write a payload into, with the timestamp, thread id, and tag
pre-loaded.  It returns NULL if tracing is disabled, the event
group is masked, or the trace buffer is full.

re-implemented ktrace_name() against ktrace_open() and ktrace()
as an inline function against ktrace_open()

Update ktracedump to handle the changes to the wire format.

Change-Id: Ifc1d58beb7ca8fc017b9a6b4d0edb97cdb908cf6