[udev][usb-xhci] Power cycle root hub ports at boot
Sometimes we never see the "changed" bits for port connection on the PORTSC register,
perhaps due to interactions with the bootloader.
This results in us sometimes missing port connections for devices that are
already connected to a root hub when we boot.
Power cycling all the root hub ports at boot ensures we start off with a clean slate.

Change-Id: I900c28858054329d476cdc4b5c8dbcf452570316