[devmgr] Fix cross-filesystem rename refcounting issue
vfs_walk returns a vnode which may or may not have been
"vn_acquire"-d. The old code for "vfs_rename" mistakenly assumed that
mount points would be acquired during path walking -- this is not
the case. Only non-mount point directories are acquired.

Comment added to clarify this behavior, and vfs_walk updated to acquire
a vnode in all success cases.

Change-Id: I9c23bfd1c0550f04f356e8fa15f0a7bb97830a23