apps/media: fix namespace issues in flog source files, etc
This CL fixes a new namespace issues in the recently-imported flog
source code. Some have to do with changes to the source tree
structure. factory_service_base* and incident* are moved from
mojo::media to mojo, because the flog service shouldn't take a
dependency on things in the mojo::media namespace. Some changes
fix outright mistakes in a previous CL.

Change-Id: Ia4115ea91eb64a9b85725464e670df694583baed