[kernel][user_copy] Move C++ bits of user_copy.h into user_ptr.h.
This is so I'll be able to add (copy-in/out) methods to user_ptr<T>
(that are always correctly typed). Without this change, it'd be
impossible to do so incrementally (at least non-hackily) since
user_copy.h includes user_ptr.h.

Hopefully we'll be able to get rid of the overloaded free functions that
take user_ptr eventually.

Note that with this change, user_copy.h becomes a pure C header (also
usable from C++).

Change-Id: I7eb22f8c2a271b82ebffe96c9caf77f3802674dc