[udev][usb] Fix teardown path for USB hubs are disconnected with devices attached
Previously the usb-bus driver recursively removed devices when a hub was disconnected.
Unfortunately this resulted in the USB host stack not being notified and was not
disabling the endpoints and slot for the devices attached to the disconnected hub.
To fix this, the usb-hub driver now handles this case from its unbind callback
and properly informs the host controller that the devices connected to the hub
were disconnected.

Change-Id: I123d4c8c31f4b35f41734dc5fc04730e06932055