[kernel][new] Fix AllocChecker
John challenged me to prove the alloc checker worked and
in doing so we realized it as flipping broken, basically
the previous tests were poor or the compiler back then
skipped running the ctor. With lot of hair pulling and cursing
we stumbled (or John did) into a solution.

Now this time a test that seems to actually excersise the
constructor.

Change-Id: I904365e1502d827b9918c6c8ef86d46d9cc55c91