[mxtl][wavl] Make WAVLTreeNodeState a standard layout.
Refactor the way in which rank is stored so the WAVLTreeNodeState<>
struct is considered to be a "standard layout".

Previously, WAVLTreeNodeState<> derived from a WAVLTreeNodeRank<> base
class which determined whether or not the rank of the node was stored
as an explicit int32_t, or simply as bool representing the rank
parity.  In operation, only the rank parity is needed (and some
operations can become slightly cheaper by recognizing this), but
during testing, we sometimes store and manipulate the explicit rank in
order to sanity check that the rank rule holds at all times.

There are many rules which a class/struct must follow in order to be
considered to have a "standard layout".  One of them is that all of
the data members of the class/struct also be of standard layout.
Another (paraphrasing here) is that all of the data members of a
class/struct exist at one level of the class hierarchy for the object.
IOW - if C derives from B which derives from A, only one of the three
may have any data members.

Previously, WAVLTreeNodeState<> held tree pointer data members and
derived from WAVLTreeNodeRank<> which held the storage for the node's
rank.  Spreading the storage across two classes in the hierarchy
caused WAVLTreeNodeState<> to be of non-standard layout.

Now, WAVLTreeNodeStateBase<> holds all of the data members as well as
common methods, and the specialization of WAVLTreeNodeState<> (which
derives from WAVLTreeNodeStateBase<>) holds the methods which
manipulate the rank in different ways.

Generally, it is nice to have a standard layout when possible as it
can make interoperation with other languages (usually C) easier in
some cases.  As an example, certain operations (such as offsetof) can
only be performed on objects with standard layouts.

Change-Id: Id37f0694ca443700cc395bab58282e81acca417e