[gfxconsole] Implement line wrapping behavior.
The cursor is allowed to exist on a position one past the right edge of
the screen such that printing the next character causes a line wrap to
occur.  This implements the "auto_right_margin" described by terminfo.

This change also fixes the backspace behavior so that it actually erases
the previous character.  Note that the "auto_left_margin" behavior for
backspace has been preserved.

Added some assertions to make it less likely we'll attempt to perform
accesses outside the bounds of the buffer.  Also fixed a couple of
off-by-one errors in the movement clamping logic which might have
caused this to happen previously.

Note that this change exposes an issue in the cursor drawing logic
where a cursor drawn into the overflow region to the right of the screen
during a line wrap is not properly erased.  (might be considered a feature)

Change-Id: I68ea02772adb1655c841175735c6a5d9287d3f63