[kernel][ihda] Stop building the kernel mode Intel HDA audio driver
This driver is in the process of being moved into user-land, the
kernel mode driver is obsolete.  Changes at the PCI level are going to
make it difficult to maintain.  Delete the makefile for now, but leave
the code checked in as a reference.  When the first revision of the
user-mode driver gets checked in, this reference code will be removed
as well.

Change-Id: Iff6efbb4fe49d107f69e571784891298f3de38c0