Remove the argument self from method Link::Watch().
Instead, watchers are never notified of changes that happen
on the same Link handle they were registered on.

If a client wants to be notified of its own changes, it can
Watch() a Dup()ed handle instead of the original.

Example that demonstrates this (and fixes LinkHost to
support watching closed handles) follows.

Change-Id: Ia91d46c140bbb0e6b63b8a283fc299b79887c8b0