Use recursively structured and typed data for Link values.
The value is a union of all types supported in JSON. If we want to
we represent JSON-LD this way. For now it just allows for typed and
structured data.

No differential update is supported yet: The Link value is always
set and transmitted as a whole. It will be easier to support deltas
once we define IDs. Anyway, I plan to represent deltas using the
same data type, so that we don't have the morass of mutation types
we used to have in the dart implementation.

The schema parameter in Link is gone for now, as it's not yet used.
I suspect we don't actually need schema information in code, only
in the component description, so maybe it won't even come back.

Change-Id: I2490f019506e3fba3018d602c2aa7346841e9623