apps/media: Add explicit EndOfStream logging message for player
The player was logging a Primed message on end-of-stream. Primed
generally implies a transition to Primed state, which is reasonable
for end-of-stream. End-of-stream also implies that the player isn't
trying to play anymore, which Prime doesn't convey. As a result, the
digest viewer, when showing the state after completed playback,
indicated a 'suspense' condition in which there was an incomplete
transition from Primed to Playing state.

This CL adds a new log message (EndOfStream) for the player
specifically for end of stream. This allows the digest to report to
be explicit about the fact that the player has reached end of stream
and to eliminate the false suspense condition.

Change-Id: I722d302360df5017049d74f3909f2878fa50b059