[bootloader] Add a boot menu
The only two options so far are selecting a framebuffer resolution and
doing a hard reset.

The framebuffer selection menu lists all the resolutions supported by
the display. After selecting a resolution, the user has 5 seconds to
confirm the choice (in order to prevent a bad framebuffer mode from
making it impossible to use the device). The cmdline param that will
select that mode by default is also displayed on the console.

Refactored much of osboot and framebuffer code to use the global xefi
symbols (gSys, gBS, etc) since we're relying on them implicitly in
printf anyway.
Change-Id: I666641f5868c8c94ee9b8c4ad03961fdd22e958e