[system] Return error for address width mismatch
Fixes mg-310.

intel_serialio_i2c_remove_slave now returns an error code if an error
occurred when trying to remove a slave.  Previously, an error code was
assigned to the status variable on condition of "Chip address width
mismatch," but the function would nevertheless ignore this variable and
return NO_ERROR instead.

Change-Id: Ia70aac31908391a89d25129717b9dd9dea3c8a24