[udev][usb-bt-hci] Don't clobber packet_size
MG-313 #done

The packet_size is no longer clobbered when the event buffer already
contains data.

Prior to this commit, the value stored to packet_size
(hci->event_buffer_packet_length) in the else case was never read, as it
was immediately clobbered by a store to packet_size on line 105. This
commit initializes packet_size to "buffer[1] + 2" (which is the value
that it gets in every case apart from the else case), and does not
subsequently write to it apart from in the else case.

Change-Id: I890faa242654b7e09fd1a19189a11f9ea5127645